# SI/02_Software_Design/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Raggy Bot API",
    "description": "API Endpoint for TRL Retrieval-Augmented Generation",
    "version": "1.0.0"
  },
  "paths": {
    "/raggy/trl": {
      "post": {
        "summary": "Process Trl Query",
        "description": "Accepts a single user query regarding TRL levels.\nProcesses the query through the LangChain RAG chain with RBAC enforcement.",
        "operationId": "process_trl_query_raggy_trl_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/internal/metadata/sessions/{session_id}": {
      "get": {
        "summary": "Get Metadata By Session",
        "operationId": "get_metadata_by_session_internal_metadata_sessions__session_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MetadataRecordListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/internal/metadata/requests": {
      "get": {
        "summary": "Get Recent Metadata Records",
        "operationId": "get_recent_metadata_records_internal_metadata_requests_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MetadataRecordListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AssessmentResultPayload": {
        "properties": {
          "candidate_level": {
            "type": "integer",
            "title": "Candidate Level"
          },
          "matched_level": {
            "type": "integer",
            "title": "Matched Level"
          },
          "decision_status": {
            "type": "string",
            "title": "Decision Status"
          },
          "reasoning_summary": {
            "type": "string",
            "title": "Reasoning Summary"
          }
        },
        "type": "object",
        "required": [
          "candidate_level",
          "matched_level",
          "decision_status",
          "reasoning_summary"
        ],
        "title": "AssessmentResultPayload"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "MetadataRecord": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp"
          },
          "response_status": {
            "type": "string",
            "title": "Response Status"
          },
          "route_path": {
            "type": "string",
            "title": "Route Path"
          },
          "model_name": {
            "type": "string",
            "title": "Model Name"
          },
          "workflow_mode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workflow Mode"
          },
          "decision_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision Status"
          }
        },
        "type": "object",
        "required": [
          "request_id",
          "user_id",
          "role",
          "timestamp",
          "response_status",
          "route_path",
          "model_name"
        ],
        "title": "MetadataRecord"
      },
      "MetadataRecordListResponse": {
        "properties": {
          "records": {
            "items": {
              "$ref": "#/components/schemas/MetadataRecord"
            },
            "type": "array",
            "title": "Records"
          }
        },
        "type": "object",
        "required": [
          "records"
        ],
        "title": "MetadataRecordListResponse"
      },
      "QueryRequest": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "candidate_level": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Candidate Level"
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "QueryRequest"
      },
      "QueryResponse": {
        "properties": {
          "answer_markdown": {
            "type": "string",
            "title": "Answer Markdown"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "mode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mode"
          },
          "assessment_result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AssessmentResultPayload"
              },
              {
                "type": "null"
              }
            ]
          },
          "missing_evidence": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Missing Evidence"
          },
          "next_question": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Question"
          }
        },
        "type": "object",
        "required": [
          "answer_markdown"
        ],
        "title": "QueryResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}